Fix : 1. 비밀번호 히스토리 3개만 유지 Fix : 2. Validation Dependency 수정
