+++ src/main/java/com/sparta/stairs/common/SwaggerConfig.java
@@ -1,28 +1,34 @@
 package com.sparta.stairs.common;
 
+
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
-import io.swagger.v3.oas.models.info.Info;
 import io.swagger.v3.oas.models.security.SecurityRequirement;
 import io.swagger.v3.oas.models.security.SecurityScheme;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import io.swagger.v3.oas.models.info.Info;
 
 @Configuration
 public class SwaggerConfig {
-
 	@Bean
-	public OpenAPI customOpenAPI() {
+	public OpenAPI openAPI() {
+		Info info = new Info()
+				.version("v1.0.0")
+				.title("News Feed 프로젝트 테스트");
+
+		String jwt = "JWT";
+		SecurityRequirement securityRequirement = new SecurityRequirement().addList(jwt);
+		Components components = new Components().addSecuritySchemes(jwt, new SecurityScheme()
+				.name(jwt)
+				.type(SecurityScheme.Type.HTTP)
+				.scheme("bearer")
+				.bearerFormat("JWT")
+		);
+
 		return new OpenAPI()
-				.addSecurityItem(new SecurityRequirement().addList("bearerAuth")) // 전역 SecurityRequirement 추가
-				.components(new Components()
-						.addSecuritySchemes("bearerAuth",
-								new SecurityScheme()
-										.type(SecurityScheme.Type.HTTP)
-										.scheme("bearer")
-										.bearerFormat("JWT")))
-				.info(new Info()
-						.title("stairs 프로젝트 테스트")
-						.version("v1.0.0"));
+				.info(info)
+				.addSecurityItem(securityRequirement)
+				.components(components);
 	}
 }
